+++ src/cljs/cxengage_javascript_sdk/modules/logging.cljs
@@ -1,5 +1,6 @@
 (ns cxengage-javascript-sdk.modules.logging
-  (:require-macros [cxengage-javascript-sdk.macros :refer [def-sdk-fn]])
+  (:require-macros [cxengage-javascript-sdk.macros :refer [def-sdk-fn]]
+                   [lumbajack.macros :refer [log]])
   (:require [cljs.core.async :as a]
             [cljs.spec :as s]
             [cxengage-javascript-sdk.domain.protocols :as pr]
@@ -7,6 +8,7 @@
             [cxengage-javascript-sdk.interop-helpers :as ih]
             [cxengage-javascript-sdk.internal-utils :as iu]
             [cxengage-javascript-sdk.domain.specs :as specs]
+            [cxengage-javascript-sdk.domain.rest-requests :as rest]
             [cxengage-javascript-sdk.state :as state]
             [cljs-uuid-utils.core :as uuid]
             [lumbajack.core :as jack]))
@@ -19,34 +21,6 @@
            :level "info"
            :message (js/JSON.stringify (ih/camelify {:data (clojure.string/join " " data) :original-client-log-level (name level)}))
            :timestamp (.toISOString date-time))))
-
-(defn log*
-  [level & data]
-  (if (some #{level} (state/get-valid-log-levels))
-    (let [level (keyword level)]
-      (doseq [d data]
-        (apply (partial jack/log* level) d))
-      (when (state/get-unsaved-logs)
-        (state/append-logs! {:data data :level level}))
-      nil)
-    nil))
-
-;; -------------------------------;;
-;; CxEngage.logging.dumpLogs();
-;; -------------------------------;;
-
-(s/def ::dump-logs-params
-  (s/keys :req-un []
-          :opt-un [::specs/callback]))
-
-(def-sdk-fn dump-logs
-  {:validation ::dump-logs-params
-   :topic-key :logs-dumped}
-  [params]
-  (let [{:keys [topic callback]} params]
-    (p/publish {:topics topic
-                :response (state/get-unsaved-logs)
-                :callback callback})))
 
 ;; -------------------------------;;
 ;; CxEngage.logging.setLevel({ level: "info" });
@@ -62,9 +36,9 @@
   [params]
   (let [{:keys [level topic callback]} params
         level (keyword level)]
-    (state/set-log-level! level jack/levels)
+    (ih/set-log-level! level)
     (p/publish {:topics topic
-                :response (state/get-log-level)
+                :response level
                 :callback callback})))
 
 ;; -------------------------------;;
@@ -81,22 +55,21 @@
   [params]
   (let [{:keys [topic callback]} params
         logs (reduce (fn [acc x] (let [log (format-request-logs x)]
-                                   (conj acc log))) [] (state/get-unsaved-logs))
-        request-map {:url (iu/api-url
-                           "tenants/:tenant-id/users/:resource-id/logs"
-                           {:tenant-id (state/get-active-tenant-id)
-                            :resource-id (state/get-active-user-id)})
-                     :method :post
-                     :body {:logs logs
-                            :device "client"
-                            :app-id (str (uuid/make-random-squuid))
-                            :app-name "CxEngage Javascript SDK"}}
-        {:keys [status api-response]} (a/<! (iu/api-request request-map true))]
+                                   (conj acc log))) [] (jack/get-unsaved-logs))
+        logs-body {:logs logs
+                   :device "client"
+                   :app-id (str (uuid/make-random-squuid))
+                   :app-name "CxEngage Agent Front-end"}
+        {:keys [status api-response]} (a/<! (rest/save-logs-request logs-body))]
     (when (= status 200)
+      (jack/wipe-logs!)
       (p/publish {:topics topic
                   :response api-response
-                  :callback callback})
-      (state/save-logs))))
+                  :callback callback}))))
+
+(defn log* [level & args]
+  (doseq [arg args]
+    (log level arg)))
 
 ;; -------------------------------------------------------------------------- ;;
 ;; SDK Logging Module
@@ -108,7 +81,11 @@
     (let [module-name :logging]
       (ih/register {:api {module-name {:set-level set-level
                                        :save-logs save-logs
-                                       :dump-logs dump-logs}}
+                                       :debug (partial log* :debug)
+                                       :info (partial log* :info)
+                                       :warn (partial log* :warn)
+                                       :error (partial log* :error)
+                                       :fatal (partial log* :fatal)}}
                     :module-name module-name})
       (ih/send-core-message {:type :module-registration-status
                              :status :success